requirements.txt
asgiref==3.2.4
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-cors-headers==3.2.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.9
oauthlib==3.1.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.1
urllib3==1.25.8
